requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
cryptography==41.0.7